refactor: refatoracao do enum de cobrança
