+++ backend/src/main/java/br/com/ufape/backend/model/Servico.java
@@ -1,5 +1,6 @@
 package br.com.ufape.backend.model;
 
+import br.com.ufape.backend.enums.FormaCobranca;
 import jakarta.persistence.*;
 
 @Entity
