+++ backend/src/test/java/br/com/ufape/backend/service/ServicoServiceTest.java
@@ -16,6 +16,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
+import br.com.ufape.backend.enums.FormaCobranca;
 
 import java.util.Optional;
